requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
python-multipart==0.0.10
requests==2.32.3
pymupdf==1.24.10
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
python-dotenv==1.0.1